requirements.txt
#chainlit==1.0.200
chainlit
langchain
langchain_community
PyPDF2
chromadb
groq
langchain-groq
ollama
python-dotenv
PyMuPDF
